Backend/requirements.txt
<<<<<<< Updated upstream
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==42.0.7
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.7
inflection==0.5.1
jwt==1.3.1
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
=======
>>>>>>> Stashed changes
